% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PresMod.R
\name{PresMod}
\alias{PresMod}
\title{Simulated Random Observation of Traits in Partially Preserved Material}
\usage{
PresMod(n, c, s, i = 1, value_spec = "count")
}
\arguments{
\item{n}{An integer specifying the number of individuals/elements in the 
population.}

\item{c}{Cases, i.e. the number or portion of individuals/elements carrying 
the trait of interest. Depending on the specification of 
\code{value_spec}, \code{c} is either given as an integer, expressing 
the absolute number of cases (\code{value_spec} = 'count') or as a 
decimal value between 0 and 1, expressing the portion of 
individuals/elements in \code{n} that are affected (\code{value_spec} = 
'portion').}

\item{s}{Sample, i.e. the number or portion of individuals/elements in the 
population that are to survive taphonomy. Depending on value 
specification (\code{value_spec}), \code{s} is either given as an 
integer, expressing the absolute number of preserved 
individuals/elements (\code{value_spec} = 'count') or as a decimal value 
between 0 and 1, expressing the portion of individuals/elements in the 
population that are preserved (\code{value_spec} = 'portion').}

\item{i}{Integer specifying the number of iterations. By default, only one 
sample is drawn.}

\item{value_spec}{The method how \code{c} and \code{s} are specified, 
either as absolute counts of cases and preserved individuals/elements or 
as portions of \code{n}. Options are the strings 'count' and 'portion'; 
the default is 'count'.}
}
\value{
A data frame is returned, giving the specified parameters \code{n}, 
   \code{c}, \code{s} and \code{i}, and the number of preserved 
   individuals/elements on which the trait of interest could be observed 
   (c_count). The rows of the data frame represent the individual iterations.
}
\description{
A given population consists of a number of individuals/elements, a certain 
number of which carry a certain trait. These are referred to as 'cases'. 
The function creates a series of random subsets from this virtual 
population by drawing a specified number of individuals/elements, 
simulating random taphonomic processes. The number of iterations (i.e. 
draws) can be specified.
}
\details{
A vector of length n is created, consisting of values '0' for individuals 
not carrying the trait of interest and '1' for cases. For each iteration, a 
sample of size \code{s} is drawn from this vector, using the function 
\code{\link[base]{sample}}. 
The number of cases in these samples (c_count) is determined using the 
function \code{\link[base]{sum}}.

The resulting data can be used for modelling the impact of material 
preservation on the calculation of frequencies. By adapting the parameters 
to a real population under study, the model can be applied to specific 
research projects.
}
\section{References}{

   Waldron T. 1991. Rates for the job: Measures of Disease Frequency in
   Paleopathology. International Journal of Osteoarchaeology 1(1):17–25.


   Waldron T. 2009. Palaeopathology. Cambridge: Cambridge University Press. 
   p. 249-261.
}

