% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PresSer.R
\name{PresSer}
\alias{PresSer}
\title{Series of Simulated Random Observations in Partially Preserved Material}
\usage{
PresSer(n, c, s, i = 1, value_spec = "count")
}
\arguments{
\item{n}{An integer or vector specifying the number(s) of 
individuals/elements in the population(s).}

\item{c}{Cases, i.e. the number or portion of individuals/elements carrying 
the trait of interest in a scenario. Depending on the specification of 
\code{value_spec}, \code{c} is either given as an integer or vector 
expressing the absolute number(s) of cases (\code{value_spec} = 'count') 
or as a decimal value between 0 and 1 or vector of such values 
expressing the portion(s) of individuals/elements in \code{n} that are 
affected (\code{value_spec} = 'portion').}

\item{s}{Sample, i.e. the number or portion of individuals/elements in the 
population that are to survive taphonomy. Depending on the specification 
of \code{value_spec}, \code{s} is either given as an integer or vector 
expressing the absolute number(s) of preserved individuals/elements 
(\code{value_spec} = 'count') or as a decimal value between 0 and 1 or 
vector of such values expressing the portion(s) of individuals/elements 
in the population that are preserved (\code{value_spec} = 'portion').}

\item{i}{Integer or vector specifying the number(s) of iterations. By 
default \code{i} is set to 1.}

\item{value_spec}{The method how \code{c} and \code{s} are specified, 
either as absolute counts of cases and preserved individuals/elements or 
as portions of \code{n}. Options are the strings 'count' and 'portion'; 
the default is 'count'.}
}
\value{
A data frame is returned, giving the output of 
   \code{\link{PresMod}} for all combinations of parameters. If 
   \code{value_spec} = 'portion', the specified portions of \code{c} and 
   \code{s} are also given ('c_portion' and 's_portion').
}
\description{
The function draws on the function \code{\link{PresMod}} to create a series 
of hypothetical preservation states from populations of defined size, with 
given numbers of individuals displaying a certain trait and a given portion 
of individuals unaffected by taphonomy.
}
\details{
The parameters, with the exceptions of \code{value_spec}, can be vectors of 
any length, and their lengths do not have to match. The function runs the 
\code{\link{PresMod}} function on all possible combinations of parameters.

The resulting data can be used for modelling the impact of material 
preservation on the calculation of frequencies under diverse conditions.
}
\section{Note}{

   Note that if \code{c} and \code{s} are specified as a portions 
   (\code{value_spec} = 'portion'), the function uses the 
   \code{\link[base]{round}} function with the rounding behaviour 
   implemented there. Please make sure that you approve of this procedure. 
   In case of doubt, it is advisable to give absolute numbers of cases and 
   preserved individuals/elements (\code{value_spec} = 'count').
}

\examples{

## Specification of case numbers and sample sizes as absolute numbers
PresSer(n=c(60, 70, 80), c=c(5, 12), s=55, i=2)

## Specification of case numbers and sample sizes as portions
PresSer(n=c(60, 70, 80), c=0.8, s=c(0.65, 0.55), i=3, value_spec="portion")

}
\seealso{
\code{\link{PresMod}}
}
