% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProportionalAgeDistribution.R
\name{ProportionalAgeDistribution}
\alias{ProportionalAgeDistribution}
\title{Proportional Distribution of Age Estimates Among Age Categories}
\usage{
ProportionalAgeDistribution(minAge = NA, maxAge = NA,
  cats = seq(floor(min(minAge)), ceiling(max(maxAge)), 1),
  accept.empty = FALSE)
}
\arguments{
\item{minAge, maxAge}{Vectors of integers specifying minimum and maximum 
age-at-death estimates for a series of individuals.}

\item{cats}{A vector of breaks defining age categories. The breaks are 
understood as points on a time line and not as units that might be part 
of one of the categories they divide. If no value is defined, single 
years within the age range defined by \code{minAge} and \code{maxAge} 
are used as categories.}

\item{accept.empty}{A boolean statement specifying whether the function 
should accept calls not providing age estimates, i.e. with \code{minAge} 
and \code{maxAge} being empty. If set to TRUE, the function will return 
a vector with value zero for all age categories. If set to FALSE (the 
default), the function will abort with an error message. Setting 
accept.empty to TRUE can be helpful in batch processing where empty 
inputs might be created by automated calculations.}
}
\value{
A vector of the same length as numbers of categories 
   (length(cats) - 1) is returned, giving the numbers of individuals that 
   have been assigned to each age category. As individuals are distributed 
   among several categories, these numbers might be fractions.
}
\description{
Age-at-death estimates from skeletons are usually expressed in age spans, 
each delimited by a minimum and a maximum age-at-death estimate, 
respectively. The function divides the contribution from each individual by 
the number of years covered by the respective age span and assigns the 
resulting fraction to each year in the span. Then, it sums up these yearly 
contributions from individuals to determine contributions from all 
individuals to specified age categories. This procedure is referred to, 
here, as 'proportional distribution of age estimates'.
}
\details{
Boldsen (1988) describes the reconstruction of mortality profiles. These 
can also be expressed as distributions of age-at-death estimates, as 
performed by this function.
}
\section{Note}{

   As \code{minAge} and \code{maxAge} refer to the same number of 
   individuals, they have to be of identical length.
}

\section{References}{

   Boldsen JL. 1988. Two Methods for Reconstructing the Empirical Mortality 
   Profile. Human Evolution 3(5):335-342.
}

