% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_Pserrabs.mod1}
\alias{sp_Pserrabs.mod1}
\title{Prevalence Calibration Model}
\format{
An R object of class \code{lm} as produced by the function 
   \code{\link[stats]{lm}}.
}
\source{
The data object was produced as part of Felix Engel's PhD thesis, 
   entitled 'Traces of Armed Conflict: Methodological Implications from 
   Anthropological Graveyard Analyses - Applied to the Site Lauchheim 
   'Wasserfurche' With Special Reference to Cranial Trauma'.
}
\usage{
sp_Pserrabs.mod1
}
\description{
This is the model which is employed by the function \code{\link{Pcal}} to 
calibrate prevalences of traits observed in partially preserved skeletal 
populations.
}
\keyword{datasets}
