% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialNeighbours.R
\name{SpatialNeighbours}
\alias{SpatialNeighbours}
\title{Nearest Neighbours Among Data Points on a Two-dimensional Map}
\usage{
SpatialNeighbours(ID, x, y, n, l, l.included = TRUE, l.x = NA, l.y = NA)
}
\arguments{
\item{ID}{A vector of distinct identifiers for all cloud points.}

\item{x, y}{Numeric vectors containing the x- and y-coordinates of the 
points specified in \code{ID}.}

\item{n}{An integer specifying the number of individuals that are to form 
the group of spatial neighbours to be determined. \code{l.included} 
specifies whether this group contains the individual defined in \code{l} 
or not.}

\item{l}{The identifier of the inidvidual for which nearest neighbours are 
to be determined. \code{l.included} specifies whether \code{l} is part 
of the group of spatial neighbours to be determined. This also 
determines whether it needs to be an element of \code{ID} or not.}

\item{l.included}{A logical operator. If set to TRUE (this is the default), 
the individual defined by \code{l} is part of the group of spatial 
neighbours to be determined. In this case, \code{l} has to be an element 
of \code{ID}. If set to FALSE, \code{l} will not be part of the 
resulting group and it does not matter if it is an element of \code{ID} 
or not.}

\item{l.x, l.y}{Numeric values specifying the x- and y-coordinates of the 
point specified by \code{l}. Specification of these parameters only 
makes sense if \code{l.included} == FALSE and \code{l} is not an element 
of \code{ID}.}
}
\value{
A vector of length \code{n} is returned, giving the distances to 
   \code{l} for all nearest neighbours that have been determined. The 
   names of the vector elements give the points identifiers from \code{ID}.
}
\description{
The function takes a range of points on a two-dimensional surface as input, 
specified by vectors of x- and y-coordinates and a vector of identifiers. 
For any point in this cloud, the function determins a specified number 
of other points with smallest distances to the point in questions. 
The calculation of distances can be carried out in a non-projected space or 
as distances between geographical coordinates.
}
\section{Note}{

   As \code{ID}, \code{x} and \code{y} all refer to the same point set, 
   they have to be of the same length.
}

